% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keboola.r.custom.application.tree.R
\docType{class}
\name{RTree-class}
\alias{RTree}
\alias{RTree-class}
\title{Application which runs KBC transformations in R}
\description{
Application which runs KBC transformations in R
}
\section{Methods}{

\describe{
\item{\code{initialize(args = NULL)}}{Constructor.
\subsection{Parameters}{\itemize{
\item{\code{args} Optional name of data directory, if not supplied then it
will be read from command line argument or environment KBC_DATADIR.}
}}
\subsection{Return Value}{Response body - either list or string in case the body cannot be parsed as JSON.}}

\item{\code{run()}}{Main application function, this should be overriden in child class.
\subsection{Return Value}{NULL}}
}}

